requirements.txt
# ── Core RAG Frameworks ───────────────────────────────────────────────────────
llama-index>=0.10.27
llama-index-embeddings-huggingface>=0.2.0
llama-index-llms-anthropic>=0.1.0
llama-index-postprocessor-cohere-rerank>=0.1.0

# ── LLM & Embeddings ──────────────────────────────────────────────────────────
anthropic>=0.40.0
sentence-transformers>=3.0.0          # BAAI/bge-large-en-v1.5 (free, local)
cohere>=5.4.0                         # Reranking (free tier: 1000 calls/month)

# ── Vector Database ───────────────────────────────────────────────────────────
chromadb>=0.5.0                       # Local persistent vector store (default)
qdrant-client>=1.9.0                  # Production cloud option

# ── Document Parsing (HTML) ───────────────────────────────────────────────────
beautifulsoup4>=4.12.0                # HTML parsing for EDGAR 10-K filings
lxml>=5.0.0                           # Fast XML/HTML parser backend for bs4
unstructured[html]>=0.14.0            # Semantic block extraction from HTML

# ── Retrieval Utilities ───────────────────────────────────────────────────────
rank_bm25>=0.2.2                      # BM25 sparse retrieval (hybrid search)

# ── Guardrails ────────────────────────────────────────────────────────────────
guardrails-ai>=0.5.0

# ── RAG Evaluation ────────────────────────────────────────────────────────────
ragas>=0.1.14
deepeval>=0.21.0

# ── Web Framework & API ───────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit>=1.36.0

# ── Data & Config ─────────────────────────────────────────────────────────────
pydantic>=2.6.0
pydantic-settings>=2.1.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.66.0
datasets>=2.18.0                      # HuggingFace datasets (evaluation)

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
